// src/app/shared/model/situacion-actual/Clues.ts
export class Clues {
    cveClues: number;
    refClues: string;
    nomClues: string;
    clue: string;
    nomEntidad: string;
    cveEntidad: string;
    nomMunicipio: string;
    cveMunicipio: string;
    nomLocalidad: string;
    cveLocalidad: string;
    nomNivelAtencion: string;
    cveNivelAtencion: number;
    desUrlSharePoint: string;
}